Move/Format/deBOM/LF/Space source code (except for tests)
