--- BlaTeX.csproj
@@ -1,31 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.BlazorWebAssembly">
-	<PropertyGroup>
-		<OutputType>Exe</OutputType>
-		<TargetFramework>net6.0</TargetFramework>
-		<LangVersion>preview</LangVersion>
-		<TypeScriptCompileBlocked>true</TypeScriptCompileBlocked>
-		<DefaultItemExcludes>$(DefaultItemExcludes);js\**</DefaultItemExcludes>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<Watch Include="Default" Exclude="../**/obj/**,**/node_modules/**" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly" Version="6.0.0-preview.6.21355.2" />
-		<PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.DevServer" Version="6.0.0-preview.6.21355.2" PrivateAssets="all" />
-		<!-- Reference JBSnorro locally in DEBUG, otherwise use Nuget package.-->
-		<PackageReference Include="JBSnorro" Version="0.0.11" Condition="'$(Configuration)'!='Debug'" />
-		<Reference Include="JBSnorro" Condition="'$(Configuration)'=='Debug'">
-			<HintPath>..\JBSnorro\JBSnorro\bin\Debug\net6.0\JBSnorro.dll</HintPath>
-		</Reference>
-	</ItemGroup>
-	<PropertyGroup>
-		<DefaultItemExcludes>$(DefaultItemExcludes);obj;Tests/**/*</DefaultItemExcludes>
-	</PropertyGroup>
-	<ItemGroup>
-		<AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
-			<_Parameter1>BlaTeX.Tests</_Parameter1>
-		</AssemblyAttribute>
-	</ItemGroup>
-</Project>
