+++ BlaTeX.sln
@@ -1,56 +1,30 @@
-﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.0.31912.275
-MinimumVisualStudioVersion = 15.0.26124.0
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BlaTeX", "BlaTeX.csproj", "{7C82F9A3-0875-4330-99B4-28A17C75766D}"
+VisualStudioVersion = 17.5.2.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlaTeX", "src\BlaTeX.csproj", "{5292DA98-CBEC-6C94-2770-0A40513C3912}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BlaTeX.Tests", "Tests\BlaTeX.Tests.csproj", "{64BCD791-9835-4726-B63D-AAEEFB2E7B5C}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{AB8D5FEC-34F4-41C4-8B47-81CCC711E12F}"
-	ProjectSection(SolutionItems) = preProject
-		.github\workflows\CI.yml = .github\workflows\CI.yml
-	EndProjectSection
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlaTeX.Tests", "Tests\BlaTeX.Tests.csproj", "{25FECCFA-F7AB-0A37-32B2-2B1E2A77D459}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
 		Release|Any CPU = Release|Any CPU
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7C82F9A3-0875-4330-99B4-28A17C75766D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7C82F9A3-0875-4330-99B4-28A17C75766D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7C82F9A3-0875-4330-99B4-28A17C75766D}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{7C82F9A3-0875-4330-99B4-28A17C75766D}.Debug|x64.Build.0 = Debug|Any CPU
-		{7C82F9A3-0875-4330-99B4-28A17C75766D}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{7C82F9A3-0875-4330-99B4-28A17C75766D}.Debug|x86.Build.0 = Debug|Any CPU
-		{7C82F9A3-0875-4330-99B4-28A17C75766D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7C82F9A3-0875-4330-99B4-28A17C75766D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7C82F9A3-0875-4330-99B4-28A17C75766D}.Release|x64.ActiveCfg = Release|Any CPU
-		{7C82F9A3-0875-4330-99B4-28A17C75766D}.Release|x64.Build.0 = Release|Any CPU
-		{7C82F9A3-0875-4330-99B4-28A17C75766D}.Release|x86.ActiveCfg = Release|Any CPU
-		{7C82F9A3-0875-4330-99B4-28A17C75766D}.Release|x86.Build.0 = Release|Any CPU
-		{64BCD791-9835-4726-B63D-AAEEFB2E7B5C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{64BCD791-9835-4726-B63D-AAEEFB2E7B5C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{64BCD791-9835-4726-B63D-AAEEFB2E7B5C}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{64BCD791-9835-4726-B63D-AAEEFB2E7B5C}.Debug|x64.Build.0 = Debug|Any CPU
-		{64BCD791-9835-4726-B63D-AAEEFB2E7B5C}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{64BCD791-9835-4726-B63D-AAEEFB2E7B5C}.Debug|x86.Build.0 = Debug|Any CPU
-		{64BCD791-9835-4726-B63D-AAEEFB2E7B5C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{64BCD791-9835-4726-B63D-AAEEFB2E7B5C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{64BCD791-9835-4726-B63D-AAEEFB2E7B5C}.Release|x64.ActiveCfg = Release|Any CPU
-		{64BCD791-9835-4726-B63D-AAEEFB2E7B5C}.Release|x64.Build.0 = Release|Any CPU
-		{64BCD791-9835-4726-B63D-AAEEFB2E7B5C}.Release|x86.ActiveCfg = Release|Any CPU
-		{64BCD791-9835-4726-B63D-AAEEFB2E7B5C}.Release|x86.Build.0 = Release|Any CPU
+		{5292DA98-CBEC-6C94-2770-0A40513C3912}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5292DA98-CBEC-6C94-2770-0A40513C3912}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5292DA98-CBEC-6C94-2770-0A40513C3912}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5292DA98-CBEC-6C94-2770-0A40513C3912}.Release|Any CPU.Build.0 = Release|Any CPU
+		{25FECCFA-F7AB-0A37-32B2-2B1E2A77D459}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{25FECCFA-F7AB-0A37-32B2-2B1E2A77D459}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{25FECCFA-F7AB-0A37-32B2-2B1E2A77D459}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{25FECCFA-F7AB-0A37-32B2-2B1E2A77D459}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {A5EF54A7-14BA-4797-9C13-EBE2F4D7E2D3}
+		SolutionGuid = {716A8873-240D-4FBA-92C5-3C7ABEBB05F5}
 	EndGlobalSection
 EndGlobal
