--- Components/Identity.razor
@@ -1,7 +1,0 @@
-﻿@* Represents the identiy component, that will just render its child content. *@
-@ChildContent
-
-@code {
-    [Parameter]
-    public RenderFragment ChildContent { get; set; }
-}
