--- Shared/MainLayout.razor
@@ -1,15 +1,0 @@
-﻿@inherits LayoutComponentBase
-
-<div class="sidebar">
-	<NavMenu />
-</div>
-
-<div class="main">
-	<div class="top-row px-4">
-		<a href="http://blazor.net" target="_blank" class="ml-md-auto">About</a>
-	</div>
-
-	<div class="content px-4">
-		@Body
-	</div>
-</div>
