+++ Tests/_Imports.razor
@@ -8,9 +8,6 @@
 @using AngleSharp.Dom
 
 @using BlaTeX
-@using BlaTeX.JSInterop
-@using BlaTeX.JSInterop.KaTeX
-@using BlaTeX.Components
 @using BlaTeX.Tests
 @using JBSnorro
 @using JBSnorro.Diagnostics
