--- _Imports.razor
@@ -1,16 +1,0 @@
-﻿@using System.Net.Http
-@using System.Net.Http.Json
-@using System.Threading.Tasks
-@using Microsoft.AspNetCore.Components.Forms
-@using Microsoft.AspNetCore.Components.Routing
-@using Microsoft.AspNetCore.Components.Web
-@using Microsoft.AspNetCore.Components.WebAssembly.Http
-@using Microsoft.JSInterop
-@using BlaTeX
-@using BlaTeX.Shared
-@using BlaTeX.JSInterop
-@using BlaTeX.JSInterop.KaTeX
-@using JBSnorro
-@using JBSnorro.Diagnostics
-@using JBSnorro.Extensions
-@using JBSnorro.Text.Json
