+++ src/App.razor
@@ -1,0 +1,14 @@
+@using BlaTeX.Shared
+@using Microsoft.AspNetCore.Components.Routing
+
+<Router AppAssembly="@typeof(BlaTeX.App).Assembly">
+    <Found Context="routeData">
+        <RouteView RouteData="@routeData"
+                   DefaultLayout="@typeof(MainLayout)" />
+    </Found>
+    <NotFound>
+        <LayoutView Layout="@typeof(MainLayout)">
+            <p>Sorry, there's nothing at this address.</p>
+        </LayoutView>
+    </NotFound>
+</Router>
