+++ src/Components/Identity.razor
@@ -1,0 +1,7 @@
+@* Represents the identiy component, that will just render its child content. *@
+@ChildContent
+
+@code {
+    [Parameter]
+    public required RenderFragment ChildContent { get; set; }
+}
