+++ src/Components/Proxy.razor
@@ -1,0 +1,19 @@
+@page "/"
+@inject HttpClient Http
+
+
+<h3>This component demonstrates calling JS from .NET. </h3>
+<input value="c = \pm \sqrt{a^2 + b^2}"
+       @oninput="onInputChanged" />
+
+<p></p>
+<KaTeX Math=@math />
+
+@code {
+    private string? math;
+
+    private void onInputChanged(ChangeEventArgs e)
+    {
+        math = (string?)e.Value;
+    }
+}
