+++ src/Services/IChildComponentMarkupService.cs
@@ -1,0 +1,11 @@
+using BlaTeX.JSInterop.KaTeX.Syntax;
+using Microsoft.AspNetCore.Components;
+
+namespace BlaTeX;
+
+public interface IChildComponentMarkupService
+{
+    IEnumerable<Range> Select(string markup);
+    RenderFragment Substitute(string markupFragment);
+    IReadOnlyList<AnyParseNode> Select(IReadOnlyList<AnyParseNode> ast);
+}
