+++ src/Shared/NavMenu.razor
@@ -1,0 +1,43 @@
+@using Microsoft.AspNetCore.Components.Routing
+
+<div class="top-row pl-4 navbar navbar-dark">
+    <a class="navbar-brand"
+       href="">BlaTeX</a>
+    <button class="navbar-toggler"
+            @onclick="ToggleNavMenu">
+        <span class="navbar-toggler-icon"></span>
+    </button>
+</div>
+
+<div class="@NavMenuCssClass"
+     @onclick="ToggleNavMenu">
+    <ul class="nav flex-column">
+        <li class="nav-item px-3">
+            <NavLink class="nav-link"
+                     href=""
+                     Match="NavLinkMatch.All">
+                <span class="oi oi-home"
+                      aria-hidden="true"></span> Proxy
+            </NavLink>
+        </li>
+        <li class="nav-item px-3">
+            <NavLink class="nav-link"
+                     href="interactive">
+                <span class="oi oi-plus"
+                      aria-hidden="true"></span> Interactive
+            </NavLink>
+        </li>
+
+    </ul>
+</div>
+
+@code {
+    private bool collapseNavMenu = true;
+
+    private string? NavMenuCssClass => collapseNavMenu ? "collapse" : null;
+
+    private void ToggleNavMenu()
+    {
+        collapseNavMenu = !collapseNavMenu;
+    }
+}
